+++ backend/nest/src/entities/index.ts
@@ -1,0 +1,5 @@
+export { User } from './user.entity';
+export { Truck } from './truck.entity';
+export { Location } from './location.entity';
+export { Favorite } from './favorite.entity';
+export { Review } from './review.entity';